requirements.txt
python-dotenv
pydub
openai
elevenlabs
# moviepy
# ipykernel